requirements.txt
langchain
langchain_core
langchain_anthropic
langchain_openai
tavily_python
langchain_community
langgraph
pyairbnb
geopy
python-dotenv
fastapi
uvicorn
jinja2
pydantic
sendgrid
numpy
pandas
tavily_python
langchain_groq
fast-flights
PyPDF2
amadeus
beautifulsoup4
googlemaps
docling
arxiv
needle-python
openai
reportlab